requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Pillow==10.0.0
exifread==3.0.0
gunicorn==20.1.0